Extract step state logic to useStepState
